pyproject.toml
[tool.poetry]
name = "softfocus-api"
version = "0.1.0"
description = ""
authors = ["João Victor <joaovictorslf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
PyMySQL = "^1.0.2"
GeoAlchemy2 = "^0.11.1"
psycopg2-binary = "^2.9.3"
bcrypt = "^3.2.2"
PyJWT = "^2.3.0"
python-decouple = "^3.6"
pydantic = {extras = ["email"], version = "^1.9.0"}
geojson-pydantic = "^0.3.4"
plpygis = "^0.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
